… xbmc/interfaces/json-rpc/schema/version.txt
… xbmc/music/MusicDatabase.cpp