Rework the way `AudioLibrary.GetArtists`, `GetAlbums` and `GetSongs` are implemented to be many times faster. 

Looking at up to a 500% increase in speed for processing JSON requests, attained by avoiding the use of fileitem lists, making inefficient queries per row for additional data and by sorting at the db rather than in memory.

Just a minor patch bump to JSON API version despite being a significant speed improvement and fix of several bugs.