@garbear 2018-07-07
Had a look at the code and it seems sane. The music database cpp got bumped by over 1,500 lines. It actually overtook VideoDatabase.cpp as the longest file. Maybe it's time to split into separate files so it can better scale to your future work?