@DaveTBlake 2018-07-11
Unless anyone else wishes to review the code, or is going to do some more JSON testing I would like to merge this Thursday so that it can be part of alpha testing in the wild.