@MilhouseVH 2018-07-11
@DaveBlake, there's a MySQL (possibly SQLite too) issue with the following JSON query:
```
curl -s --data-binary '{"jsonrpc": "2.0", "params": {"sort": {"order": "ascending", "method": "title"}, "properties": ["title", "track", "artist", "album", "fanart", "thumbnail", "file"], "limits": {"start": 0, "end": 400}}, "method": "AudioLibrary.GetSongs", "id": "libSongs"}' -H 'content-type: application/json;' http://localhost:8080/jsonrpc
```

In kodi.log:
```
00:43:07.105 T:1359635312   DEBUG: CWebServer[8080]: request received for /jsonrpc
00:43:07.118 T:1359635312   DEBUG: GetSongsByWhereJSON query: SELECT sv.*, song_artist.idArtist AS idArtist, song_artist.strArtist AS strArtist FROM (SELECT song.idSong, strTitle, strTitle, path.strPath || strFilename AS strPathFile, album.strAlbum AS strAlbum, (iTrack & 0xffff) AS track FROM song JOIN album ON album.idAlbum = song.idAlbum JOIN path ON path.idPath = song.idPath ORDER BY strTitle LIMIT 400) AS sv JOIN song_artist ON song_artist.idSong = sv.idSong AND song_artist.idRole = 1 GROUP BY sv.idSong, song_artist.idArtist ORDER BY strTitle, song_artist.iOrder
00:43:07.120 T:1359635312   ERROR: SQL: [MyMusic72] Undefined MySQL error: Code (1060)
                                            Query: SELECT sv.*, song_artist.idArtist AS idArtist, song_artist.strArtist AS strArtist FROM (SELECT song.idSong, strTitle, strTitle, path.strPath || strFilename AS strPathFile, album.strAlbum AS strAlbum, (iTrack & 0xffff) AS track FROM song JOIN album ON album.idAlbum = song.idAlbum JOIN path ON path.idPath = song.idPath ORDER BY strTitle LIMIT 400) AS sv JOIN song_artist ON song_artist.idSong = sv.idSong AND song_artist.idRole = 1 GROUP BY sv.idSong, song_artist.idArtist ORDER BY strTitle, song_artist.iOrder
00:43:07.120 T:1359635312   ERROR: GetSongsByWhereJSON failed
```

Reformatting the query produces:
```
SELECT sv.*, song_artist.idArtist AS idArtist, song_artist.strArtist AS strArtist
FROM (SELECT song.idSong, strTitle, strTitle, path.strPath || strFilename AS strPathFile, album.strAlbum AS strAlbum, (iTrack & 0xffff) AS track
      FROM song JOIN album ON album.idAlbum = song.idAlbum
                JOIN path ON path.idPath = song.idPath
      ORDER BY strTitle LIMIT 400) AS sv JOIN song_artist ON song_artist.idSong = sv.idSong AND song_artist.idRole = 1
GROUP BY sv.idSong, song_artist.idArtist
ORDER BY strTitle, song_artist.iOrder
```

Running this query in MySQL Workbench produces the error:
```
Error Code: 1060. Duplicate column name 'strTitle'
```

and indeed, in the `sv` sub query there are two references to the `strTitle` column - presumably one of them needs to be dropped?