@DaveTBlake 2018-07-12
Thanks @MilhouseVH, while SQLite happily gets same field twice obviously MySQL does not. Fix imminent.