@zag2me 2018-07-12
Tested in millhouse raspberry pi build and seems to work as expected with the iPhone remote as far as I can tell. Hard to judge the speed but any improvement is most welcome in that area!