@a1rwulf 2018-07-12
https://github.com/DaveTBlake/xbmc/blob/c5e14e91bddc077250b0b8272162fdaf0deb1da9/xbmc/music/MusicDatabase.cpp#L5802
Imho this is not good - you tie the json schema to the db schema.
They are different subsystems and MusicDatabase.cpp should not need a change when somebody does changes in the json api.