@MilhouseVH 2018-07-15
Thanks @DaveTBlake  - your latest updates now produce 100% identical results when compared with a build without this PR, ie. LibreELEC build `#0707` (methods tested: `GetArtists`, `GetAlbums`, `GetSongs`).

I know you're keen to merge this, it's probably the best way to shake out any remaining issues (hopefully there aren't any).

----

Performance comparison using MySQL 5.5.21 on HP N36L

```
Albums     : 351
Artists    : 1537
Songs      : 7384
```

Old/New timings are in seconds, averaged over 3 runs each.

JSON data retrieved in chunks of 400 items.

Tested using `texturecache.py`, eg. `texturecache.py j artists >/dev/null`

#### RPi1 (512MB RAM, 10MB/s ethernet):

| Method | Old | New |
| ---- | ---- | ---- |
| `AudioLibrary.GetArtists` | 22.615 | 5.046 |
| `AudioLibrary.GetAlbums` | 7.055 | 4.531 |
| `AudioLibrary.GetSongs` | 218.391 | 50.844 |

#### RPi3+ (1GB RAM, 30MB/s ethernet):

| Method | Old | New |
| ---- | ---- | ---- |
| `AudioLibrary.GetArtists` | 3.865 | 1.340 |
| `AudioLibrary.GetAlbums` | 1.204 | 1.436 |
| `AudioLibrary.GetSongs` | 42.823 | 15.588 |

#### Skylake i5 NUC (8GB RAM, 30MB/s 5GHz WiFi):

| Method | Old | New |
| ---- | ---- | ---- |
| `AudioLibrary.GetArtists` | 8.410 | 0.421 |
| `AudioLibrary.GetAlbums` | 1.318 | 1.036 |
| `AudioLibrary.GetSongs` | 39.441 | 22.384 |